Add assertions for compile-time errors
Fixes #36. Previously, it was not possible to write assertions that a
given program does not compile. Now, users can call the
`compileErrors(String): String` macro which takes a string literal as an
argument and returns the compile-time errors as strings (or
empty string if there is no error)

The macro is implemented for both Scala 2.x and Dotty.