Relax `assertEquals` strictness about type equality.
Previously, `assertEquals` only allowed comparing values of the same
type. Now, it's OK as long as the second argument is a subtype of the
first argument. The reasoning for this change is that it's common that
you want to test some generic value like `Option[Int]` obtained from a
method we're testing against an expected hardcoded value like
`Some[Int]`.