Fix #138 - duplicate names with bracket characters
A bug in the JUnit/sbt event handler made it ignore tests when two tests
had the same name AND the name contains a `[` character. This bug is a
relic from the original sbt/junit-interface implementation that assumes
that test names are JVM method names where `[` has special meaning.
This bug was not present in the Scala.js implementation.