Switch to Scala 3 universal equality for `assertEquals()`
Previously, MUnit had a subtyping constraint on `assertEquals(a, b)` so
that it would fail to compile if `a` was not a subtype of `b`. This was
a suboptimal solution because the compile error messages could become
cryptic in some cases. Now, MUnit uses the new `Eql[A, B]` universal
equality in Scala 3. For Scala 2, we drop the subtyping constraint so
that it's no longer a compile error to compare unrelated types. If the
values are different, then users will see a runtime error instead.