Move `Fixture[T]` into a toplevel class.
Previously, `Fixture[T]` was declared as an inner class of `FunSuite`.
This made it annoying to implement fixtures as normal toplevel classes.
Now, `Fixture[T]` is just a normal toplevel class making it easier to
reason about.

Breaking change: In order to implement this change, we remove the
ability to extend `Suite` with a custom `type TestValue`. The test value
is now hardcoded to `Future[Any]`. I'm not aware of any library that
extends `munit.Suite` with a custom `TestValue`. All the 3rdparty
integrations that I have seen hardcode against `munit.FunSuite` anyways.
Given this feature doesn't seem to be used anywhere and it complicates
the public API, I feel it's best to remove it.