Enable mima
Previously, binary compatibility was maintained on a best-effort basis.
With growing number of external libraries that integrate with MUnit,
it's critical that we take binary breaking changes seriously. This
commit enables Mima so that the CI fails when we introduce binary
breaking changes. The binary breaking changes that have happened since
v0.7.0 are mostly irrelevant for most users (internal APIs that were
accidentally public)