Attempt 1: make sure to always flush buffered logs
Previously, buffered logs would sometimes not get flushed resulting in
lost test report output. See #408. This commit adds an extra attempt to
flush the buffered logs inside the sbt testing framework in case the
logs don't get flushed by the JUnit test listener. I'm not 100% sure
this fixes the problem but it seems worth a short and it shouldn't cause
a regression.