Introduce `AnyFixture` to preserve `Unit` result type on `Fixture`
Previously, the `Fixture` lifecycle methods returned `Any`, which meant
that auto-completions would insert `Any` in the result type when
auto-generating implementation of a fixture. Now, the `Fixture[T]` class
has `Unit` as result types for consistency with older versions of MUnit.
Instead, we introduce `AnyFixture[T]` which only needs to be used when
comibing a list of `Fixture[T]` and `FutureFixture[T]` (along with
    3rdparty fixture classes).