Make flaky tag work with Scalacheck suites (#478)
* Make flaky tag work with Scalacheck suites

Previously, the Scalacheck suite transform was registered after the
`.flaky` transform so a flaky test would fail even if it was OK to
ignore flaky failures. Reordering the transforms fixes the issue.

* Fix duplicate reporting of ignored tests in JS/Native

There was a bug in the JS/Native implementation of `RunNotifier` so that
it would report that a test was both ignored and that it succeeded.