Revive "Strict Equality" for `assertEquals()` (#521)
* Make FailException and ComparisonFailException work more similarly.

Previously, FailException had some custom nice-to-have features that
ComparisonFailException didn't have.

* Introduce "strict equality" mode for `assertEquals()` and friends.

Previously, MUnit had a subtyping constraint on `assertEquals(a, b)` so
that it would fail to compile if `a` was not a subtype of `b`. This was
a suboptimal solution because the compile error messages could become
cryptic in some cases. Additionally, this API didn't integrate with
other libaries like Cats that has its own `cats.Eq[A,B]` type-class.

Now, MUnit uses a new `munit.Compare[A,B]` type-class for comparing
values of different types. By default, MUnit provides a "universal"
instance that permits comparison between all types and uses the built-in
`==` method. Users can optionally enable "strict equality" by adding the
compiler option `"-Xmacro-settings.munit.strictEquality"` in Scala 2.
In Scala 3, we use the `Eql[A, B]` type-classes instead to determine
type equality.

* Address review feedback

* Drop strict equality, allow comparison between supertypes/subtypes

This is a fourth attempt at improving strict equality in MUnit
`assertEquals()` assertions.

* First attempt (current release version): require second argument to be
  a supertype of the first argument. This has the flaw that the compile
  error message is cryptic and that the ordering of the arguments affects
  compilation.
* Second attempt: use `Eql[A, B]` in Scala 3 and allow comparing any
  types in Scala 2. This has the flaw that it's a regression in some
  cases for Scala 2 users and that `Eql[A, B]` is not really usable
  in its current form, see related discussion
  https://contributors.scala-lang.org/t/should-multiversal-equality-provide-default-eql-instances/4574
* Third attempt: implement "strict equality" for Scala 2 with a macro
  and `Eql[T, T]` in Scala. This improves the situation for Scala 2,
  but would mean relying on a feature that we can't easily port to Scala 3.
* Fourth attempt (this commit): improve the first attempt (current
  release) by allowing `Compare[A, B]` as long
  as `A <:< B` OR `B <:< A`. This is possible thanks to an observation
  by Gabriele Petronella that it's possible to layer the implicits to
  avoid diverging implicit search.

The benefit of the fourth approach is that it works the same way for
Scala 3 and Scala 3. It's very nice that we can avoid macros as well.

* Address review feedback

* Run scalafmtSbt

* Remove unused import

* Fix dotty tests in AssertionsSuite

The Scala 3 (dotty) tests now use compareSubtypeWithSupertype instead
of compareSupertypeWithSubtype. Additionally, the "unrelated" test was
not seeing the context code above and so I've moved all the code into
compileErrors.

* Add mima exclusions for assertEquals and co

* Remove unused import in scala-3 MacroCompat

* Reintroduce special-case msgs for comparing arrays

* Reintroduce better string inequality error msgs

* Update Clue deprecation to 1.0

Co-authored-by: Ólafur Páll Geirsson <olafurpg@gmail.com>

* Fix typo in AssertionsSuite test name

Co-authored-by: Ólafur Páll Geirsson <olafurpg@gmail.com>

Co-authored-by: Olafur Pall Geirsson <lgeirsson@twitter.com>
Co-authored-by: Ólafur Páll Geirsson <olafurpg@gmail.com>